move lock code into it's own file
